requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cryptography==44.0.1
Django==5.1.6
django-allauth==65.4.1
django-axes==7.0.2
django-ckeditor-5==0.2.17
django-cleanup==9.0.0
django-crispy-forms==2.3
django-js-asset==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
django-taggit==6.1.0
django-tinymce==4.1.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
